requirements.txt
psycopg2-binary==2.9.7
pymongo==4.5.0
ollama==0.2.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
setuptools